+++ tests/file/CollectionTest.php
@@ -134,16 +134,16 @@ class CollectionTest extends TestCase
             $bytes = 'Test file content ' . $i;
             $collection->insertFileContent($bytes, [
                 'index' => $i
-            ]);
+            ], ['chunkSize' => 15]);
         }
 
         $this->assertEquals(1, $collection->remove(['index' => ['$in' =>[1, 2, 3]]], ['limit' => 1]));
         $this->assertEquals(9, $collection->count());
-        $this->assertEquals(9, $collection->getChunkCollection()->count());
+        $this->assertEquals(18, $collection->getChunkCollection()->count());
 
         $this->assertEquals(3, $collection->remove(['index' => ['$in' =>[5, 7, 9]]]));
         $this->assertEquals(6, $collection->count());
-        $this->assertEquals(6, $collection->getChunkCollection()->count());
+        $this->assertEquals(12, $collection->getChunkCollection()->count());
 
         $this->assertEquals(6, $collection->remove());
         $this->assertEquals(0, $collection->count());